+++ apps/eas-expo-go/scripts/eas-build-post-install.sh
@@ -8,16 +8,16 @@ export PATH="$ROOT_DIR/bin:$PATH"
 maybe_prebuild_hermes() {
   ANDROID_DIR="$ROOT_DIR/android"
   CURRENT_VERSION=$(test -f $ANDROID_DIR/prebuiltHermes/.hermesversion && cat $ANDROID_DIR/prebuiltHermes/.hermesversion) || true
-  TARGET_VERSION=$(test -f $ROOT_DIR/react-native-lab/react-native/sdks/.hermesversion && cat $ROOT_DIR/react-native-lab/react-native/sdks/.hermesversion) || true
+  TARGET_VERSION=$(test -f $ROOT_DIR/react-native-lab/react-native/packages/react-native/sdks/.hermesversion && cat $ROOT_DIR/react-native-lab/react-native/packages/react-native/sdks/.hermesversion) || true
   if [[ $CURRENT_VERSION != $TARGET_VERSION ]]; then
     pushd $ANDROID_DIR
     rm -rf $ANDROID_DIR/prebuiltHermes
     mkdir -p $ANDROID_DIR/prebuiltHermes
-    cp -f $ROOT_DIR/react-native-lab/react-native/sdks/.hermesversion $ANDROID_DIR/prebuiltHermes/
-    ./gradlew :ReactAndroid:hermes-engine:assembleRelease
-    ./gradlew :ReactAndroid:hermes-engine:assembleDebug
-    cp -f $ROOT_DIR/react-native-lab/react-native/ReactAndroid/hermes-engine/build/outputs/aar/hermes-engine-release.aar $ANDROID_DIR/prebuiltHermes/
-    cp -f $ROOT_DIR/react-native-lab/react-native/ReactAndroid/hermes-engine/build/outputs/aar/hermes-engine-debug.aar $ANDROID_DIR/prebuiltHermes/
+    cp -f $ROOT_DIR/react-native-lab/react-native/packages/react-native/sdks/.hermesversion $ANDROID_DIR/prebuiltHermes/
+    ./gradlew :packages:react-native:ReactAndroid:hermes-engine:assembleRelease
+    ./gradlew :packages:react-native:ReactAndroid:hermes-engine:assembleDebug
+    cp -f $ROOT_DIR/react-native-lab/react-native/packages/react-native/ReactAndroid/hermes-engine/build/outputs/aar/hermes-engine-release.aar $ANDROID_DIR/prebuiltHermes/
+    cp -f $ROOT_DIR/react-native-lab/react-native/packages/react-native/ReactAndroid/hermes-engine/build/outputs/aar/hermes-engine-debug.aar $ANDROID_DIR/prebuiltHermes/
     popd
   fi
 }