+++ packages/babel-preset-expo/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### 💡 Others
 
+### 📚 3rd party library updates
+
+- Update `metro-react-native-babel-preset` to 0.76.5. ([#22588](https://github.com/expo/expo/pull/22588) by [@kudo](https://github.com/kudo))
+
 ## 9.4.0 — 2023-05-08
 
 _This version does not introduce any user-facing changes._